backend/message-passing/requirements.txt
functions-framework==3.3.0
google-cloud-pubsub==2.11.0
google-cloud-firestore==2.9.0